// paper-assistant-frontend/src/api/generated/types.gen.ts
export type ProjectDto = {
    /**
     * 项目ID
     */
    id?: string;
    /**
     * 项目名称
     */
    name?: string;
    /**
     * 项目描述
     */
    description?: string;
    /**
     * 研究方向
     */
    topic?: string;
    /**
     * 项目状态: CREATED(已创建)/SEARCHING(检索中)/SEARCHED(检索完成)/ANALYZING(分析中)/ANALYZED(分析完成)
     */
    status?: string;
    /**
     * 基础论文ID
     */
    basePaperId?: string;
    /**
     * 基础论文标题
     */
    basePaperTitle?: string;
    /**
     * 论文数量
     */
    paperCount?: number;
    /**
     * 创建时间
     */
    createTime?: string;
    /**
     * 更新时间
     */
    updateTime?: string;
};